+++ src/lib/djerba/R_plots/biomarkers_plot.R
@@ -38,7 +38,20 @@ if(biomarker=="tmb"){
   #subset tcga data to cancer type
   tcga_tmb_data_type <- tcga_tmb_data %>% filter(if (sample_tcga %in% tcga_tmb_data$CANCER.TYPE) CANCER.TYPE == sample_tcga else NA)
   
-  median_tmb <- median(tcga_tmb_data$tmb)
+  
+  if (sample_tcga %in% external_tmb_data_type$CANCER.TYPE){
+    median_tmb <- median(external_tmb_data_type$tmb)
+    cohort_label <- paste(toupper(sample_tcga) ,"Cohort")
+  }
+  else if (sample_tcga %in% tcga_tmb_data_type$CANCER.TYPE){
+    median_tmb <- median(tcga_tmb_data_type$tmb)
+    cohort_label <- paste("TCGA",toupper(sample_tcga),"Cohort")
+  }
+  else{
+    median_tmb <- median(tcga_tmb_data$tmb)
+    cohort_label <- "All TCGA "
+}
+  
   tmb_path <- paste(work_dir, 'tmb.svg', sep='/')
   
   svg(tmb_path, width = 8, height = 1.6, bg = "transparent")
@@ -46,12 +59,12 @@ if(biomarker=="tmb"){
   ggplot(tcga_tmb_data) + 
     {
       if (sample_tcga %in% external_tmb_data_type$CANCER.TYPE)
-        geom_boxplot(data = external_tmb_data_type, aes(x=0,y=tmb,color="Cohort"),width = 0.05, outlier.shape = NA) 
+        geom_boxplot(data = external_tmb_data_type, aes(x=0,y=tmb,color="Cohort"),width = 0.1, outlier.shape = NA) 
         
       else if (sample_tcga %in% tcga_tmb_data_type$CANCER.TYPE)
-        geom_boxplot(data = tcga_tmb_data_type, aes(x=0,y=tmb,color="Cohort"),width = 0.05, outlier.shape = NA) 
+        geom_boxplot(data = tcga_tmb_data_type, aes(x=0,y=tmb,color="Cohort"), width = 0.1, outlier.shape = NA) 
       else
-        geom_boxplot(aes(x=0,y=tmb,color="All TCGA"),width = 0.05, outlier.shape = NA) 
+        geom_boxplot(aes(x=0,y=tmb,color="All TCGA"),width = 0.1, outlier.shape = NA) 
     } +
   #  geom_hline(yintercept = 1,alpha=0.25,color="white")  +
   #  geom_hline(yintercept = max(sampleTMB, 25), alpha=0.25,color="white")  +
@@ -59,25 +72,25 @@ if(biomarker=="tmb"){
     annotate( geom="segment", x = -0.1, xend=0.1, y=10, yend=10, colour = "gray") +
     
     annotate(geom="text",y = 10,x=0,color="gray30",label="TMB-H Cutoff",  vjust = -4.5, size=4) +
-    annotate(geom="text",y = median_tmb, x=0,color="black",label="Cohort", hjust = 0.3, vjust = 3, size=4) +
+    annotate(geom="text",y = median_tmb, x=0,color="black",label=cohort_label, hjust = 0.5, vjust = 2.8, size=4) +
     annotate(geom="text",y = sampleTMB,x=0,color="red",label="This Sample",  vjust = -2.5,size=4) +
     
-    annotate(geom="point",y = sampleTMB,x=0,color="red",shape=1, size=5) +
-    annotate(geom="point",y = sampleTMB,x=0,color="red",shape=20, size=1.5) +
+    annotate(geom="point",y = sampleTMB,x=0,color="red",shape=1, size=8) +
+    annotate(geom="point",y = sampleTMB,x=0,color="red",shape=20, size=3) +
     
-    labs(x="",y="coding mutations/mb",color="",title="",shape="",size="") +
+    labs(x="",y="coding mutations per Mb",color="",title="",shape="",size="") +
     scale_color_manual( values= c( "gray30", "red") ) +
     scale_shape_manual(values=c(16,1)) +
     theme_classic() +
     guides(shape="none",size="none",color="none") + 
-    scale_y_continuous( limits = c(0, max(sampleTMB, 25))) +
+    scale_y_continuous( limits = c(0, max(sampleTMB, 15))) +
     coord_flip(clip = "off") +
     theme(
       axis.line.y = element_blank(),
       panel.grid = element_blank(), 
       text = element_text(size = 18),
       legend.title=element_blank(),
-      plot.margin = unit(c(0, 12, 0, 4), "points"),
+      plot.margin = unit(c(t=0, r=6, b=0, l=-20), "points"),
       axis.title.y=element_blank(),
       axis.text.y=element_blank(),
       axis.ticks.y=element_blank(),
@@ -97,42 +110,50 @@ if(biomarker=="msi"){
   
   boot <- read.table(msi_path,header=FALSE)
   
-  names(boot) <- c("q0","q1","median","q3","q4")
+  names(boot) <- c("q0","q1","median_value","q3","q4")
   boot$Sample <- "Sample"
   
   msi_out_path <- paste(work_dir, 'msi.svg', sep='/')
+  
+  msi_median <- as.numeric(unique(boot$median_value))
   
   options(bitmapType='cairo')
   svg(msi_out_path, width = 8, height = 1.6, bg = "transparent")
   print(
     
   ggplot(boot,aes(x="Sample")) + 
-    geom_errorbar(aes(ymin=q1, ymax=q3), width=0, linewidth=2) +
+    geom_errorbar(aes(ymin=as.numeric(q1), ymax=as.numeric(q3)), width=0, linewidth=1, color="red") +
     
     annotate(x = 0, xend=2, y=cutoff_MSS, yend=cutoff_MSS,geom="segment",colour = "gray") +
-    annotate(geom="text",x = 0,y=0,color="gray30",label="MSS", hjust = 0, vjust = -4.1,size=4) +
+    annotate(geom="text",x = 0,y=cutoff_MSS/2,color="gray30",label="MSS", hjust = 0.5, vjust = -6,size=4) +
     
     annotate(x = 0, xend=2, y=cutoff_MSI, yend=cutoff_MSI,geom="segment", colour = "gray") +
-    annotate(geom="text",x = 0,y=cutoff_MSI,color="gray30",label="MSI", hjust = -0.2, vjust = -4.1,size=4) +
+    annotate(geom="text",x = 0,y=(cutoff_MSI + max(msi_median, 40))/2, color="gray30",label="MSI", hjust = 0.5, vjust = -6,size=4) +
     
-    geom_bar(aes(y=median),fill='gray',stat ="identity",alpha=0.5,colour="red") + 
-    geom_errorbar(aes(ymin=q0, ymax=q4), width=0,colour="red") +
+    annotate(x = 0, xend=2, y=cutoff_MSI, yend=cutoff_MSI,geom="segment", colour = "gray") +
+    annotate(geom="text",x = 0,y=(cutoff_MSI + cutoff_MSS)/2, color="gray30",label="Inconclusive", hjust = 0.5, vjust = -6,size=4) +
     
+    
+   # geom_bar(aes(y=median),fill='gray',stat ="identity",alpha=0.5,colour="red") + 
+    annotate(geom="point",y = msi_median, x="Sample",color="red",shape=1, size=8) +
+    annotate(geom="point",y = msi_median, x="Sample",color="red",shape=20, size=3) +
     
     theme_classic() + 
     labs(x="",y="unstable microsatellites (%)",title="") + 
-    scale_y_continuous(expand = c(0,0), limit = c(0, 100)) + 
+    scale_y_continuous( limit = c(0, max(msi_median, 40))) + 
     guides(fill="none", alpha="none")+
     coord_flip() +
 
     scale_color_manual(values=c("#65bc45","#000000","#0099ad")) +
     theme(
+          axis.line.y = element_blank(),
+          legend.title=element_blank(),
           axis.title.y=element_blank(),
           axis.text.y=element_blank(),
           axis.ticks.y=element_blank(),
           text = element_text(size = 18),
           panel.grid = element_blank(), 
-          plot.margin = unit(c(0, 12, 0, 4), "points"),
+          plot.margin = unit(c(t=-20, r=-20, b=0, l=-20), "points"),
           line = element_blank(),
           panel.background = element_rect(fill = "transparent", colour = NA),
           plot.background = element_rect(fill="transparent",color=NA)
@@ -143,4 +164,7 @@ if(biomarker=="msi"){
   dev.off()
   
 }
+
+
+